requirements.txt
pefile
watchdog
psutil
plyer
ttkbootstrap
requests
packaging
gunicorn 
flask
pymongo
pyjwt
werkzeug
discord.py
